Website/Auth/views.py
from django.shortcuts import render, redirect
from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
from django.views.generic.edit import CreateView

from django.contrib.auth import login
from django.contrib.auth.decorators import login_required
from social_django.utils import psa


# https://www.youtube.com/watch?v=drMr9B3ZcPI

def login_view(request):
    if request.method == 'POST':
        form = AuthenticationForm(data=request.POST)
        if form.is_valid():
            user = form.get_user()
            login(request, user)
            request.session['token'] = user.auth_token.key
            return redirect('/')
    else:
        form = AuthenticationForm()
    return render(request, 'login.html', {'form': form})

@login_required
def home_view(request):
    return render(request, 'registration.html')