requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
defusedxml==0.6.0
Django==3.0.3
django-bootstrap4==1.1.1
django-braces==1.14.0
django-controlcenter==0.2.9
django-crispy-forms==1.8.1
django-oauth-toolkit==1.2.0
django-pkgconf==0.3.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
kiwisolver==1.1.0
matplotlib==3.2.0
mysql==0.0.2
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.1.0
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rest-social-auth==3.0.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
